requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Multi-tenancy
django-tenants==3.6.1
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
django-rest-auth==0.9.5

# AI/ML Libraries
transformers==4.36.2
torch==2.1.2
tensorflow==2.15.0
scikit-learn==1.4.0
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
langchain==0.1.0
openai==1.7.2

# Sentiment Analysis
vaderSentiment==3.3.2
flair==0.13.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Celery & Redis (for async tasks)
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Database & Storage
dj-database-url==2.1.0
whitenoise==6.6.0
boto3==1.34.21
django-storages==1.14.2

# API Documentation
drf-spectacular==0.27.0
drf-yasg==1.21.7

# Monitoring & Logging
sentry-sdk==1.39.2
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0

# Utilities
python-decouple==3.8
pillow==10.2.0
requests==2.31.0
python-dateutil==2.8.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Production
gunicorn==21.2.0

